Split validateSubmission function to validator.js
